readd fs to rollup's sys-module-plugin
we readded fs, then ran the built stencil binary in a stencil project
that only used the DCE OT. running into an issue where it hits `catch`
block in `bundle-output#bundleOptions`, which indicates that we are
throwing from `node-resolve-plugin` from one of 2 places in that.

today we tried to hack on the minified output, with some/minimal
success.