refactor(utils): add a little bit of type safety around JSON.stringify
This adds two things, the first is a type, `JsonSafe`, which takes a
type parameter and will basically evaluate (at the type level) to
`never` if the supplied type can't be serialized to JSON. This allows us
to add constraints throughout the codebase wherever we want to assert
that a value must be JSON serializable (or, well, that the type
annotated on or inferred for a value is serializable).

Then this type is used to implement a type-safe `JSON.stringify`. This
is exported from a new `utils/` module, `src/utils/json.ts`. If you call
it with type which can't be safely (meaning losslessly) serialized to
JSON you'll get a type error.