chore(release): override angular issue prefixes
this commit introduces a configuration file for conventional changelog,
in order to override the angular presets that stencil uses today. the
motivation for this change is that when a stencil developer cross
references an issue in stencil in a commit issue, conventional-changelog
+ the angular preset assumes it's an issue number that's been closed.
when writing the changelog, text "closes #ISSUE_NUMBER" is erroneously
added to CHANGELOG.md, which has to be manually fixed at release time.

doing this helps in the immediate future (the engineer performing the
release doesn't have to manually update CHANGELOG.md) and will be useful
when the team enables ci-based releases.